done user 8/12/2024

+++ ZestyBiteWebAppSolution/Repositories/Interfaces/IAccountRepository.cs
@@ -1,13 +1,11 @@
 ﻿using ZestyBiteWebAppSolution.Models.Entities;
 
 
-namespace ZestyBiteWebAppSolution.Repositories.Interfaces
-{
-    public interface IAccountRepository : IRepository<Account>
-    {
+namespace ZestyBiteWebAppSolution.Repositories.Interfaces {
+    public interface IAccountRepository : IRepository<Account> {
         Task<Account?> GetAccountByUsnAsync(string usn);
         Task<Account?> GetAccountByEmailAsync(string email);
         Task<IEnumerable<Account?>> SearchAccountByNamesAsync(string name);
-        Task<Account> CreateAccountAsync(Account account , sbyte roleId);   //  manager only
+        Task<Account> CreateAccountAsync(Account account, int roleId);   //  manager only
     }
 }
